+++ normalize_scalar_flux_kernel.c
@@ -7,5 +7,5 @@ void normalize_scalar_flux_kernel(Parameters P, float * new_scalar_flux, int cel
   if( energy_group >= P.n_energy_groups )
     return;
 
-  new_scalar_flux[cell * P.n_energy_groups + energy_group] *= P.inverse_total_track_length;
+  new_scalar_flux[(uint64_t) cell * P.n_energy_groups + energy_group] *= P.inverse_total_track_length;
 }
